# WIP: new QDense design