--- a/itensor/indexset.cc
+++ b/itensor/indexset.cc
@@ -677,4 +677,15 @@ iterInds(IndexSet const& is)
     return detail::IndexValIter(is);
     }
 
+QN
+flux(std::vector<IndexVal> const& ivs)
+    {
+    QN elt_flux;
+    for(auto const& iv : ivs)
+        {
+        elt_flux += dir(iv)*qn(iv);
+        }
+    return elt_flux;
+    }
+
 } //namespace itensor