Here is an initial prototype for a new QDense storage design and contraction algorithm. For a 4x4 Hubbard model at U=2 with particle number, spin and momentum around the cylinder conserved, the timings are the following:

This pull request
```
    vN Entropy at center bond b=8 = 3.019814122626
    Eigs at center bond b=8: 0.1677 0.1677 0.1597 0.1597 0.0138 0.0138 0.0138 0.0138 0.0096 0.0096 
    Largest link dim during sweep 5/5 was 7341
    Largest truncation error: 4.5789e-09
    Energy after sweep 5/5 is -18.876818002901
    Sweep 5/5 CPU time = 18m, 25s (Wall time = 2m, 26.5s)
```

Current v3
```
    vN Entropy at center bond b=8 = 3.019814122626
    Eigs at center bond b=8: 0.1677 0.1677 0.1597 0.1597 0.0138 0.0138 0.0138 0.0138 0.0096 0.0096 
    Largest link dim during sweep 5/5 was 7341
    Largest truncation error: 4.5789e-09
    Energy after sweep 5/5 is -18.876818002901
    Sweep 5/5 CPU time = 29m, 42s (Wall time = 7m, 52.9s)
```

The major speedups are from an improved algorithm for finding the output blocks from the input blocks (instead of from the flux), caching the information about which blocks contract to which blocks when performing the contraction, as well as storing the cartesian indices of the block locations instead of the linear indices to avoid unnecessary integer divisions.

Here are some things that need to be done before merging:

- [x] Complete the write-to-disk and read-from-disk functionality for the new QDense storage
- [ ] Decide how to deal with backwards compatibility of saved files.
- [x] Write the code for setting elements of a "block deficient" QDense storage, i.e. setting the element of a block that could exist according to the flux but doesn't because it resulted from a QDense*QDense contraction that didn't create that block.
- [x] Make sure that functions like `.fill` for QDense storage fills any blocks that should exist according to the QN flux of the tensor (unfortunately, this would slow down all calls to `.fill` to some extent, since it will need to check if all of the blocks exist).
- [x] Optimize QDense+QDense (for example, right now the logic of merging the blocks doesn't use the fact that they are already sorted, and sorting two sorted blocks can be optimized).

Thanks a lot to Alex Wietek for pointing out the performance issue when many QNs are used and helping with profiling to pin down slow parts of the code!