@mtfishman 2019-11-14
I just added functionality for creating Dense and QDense storage with undefined data, which speeds up the constructors a lot. The notation is:
```C++
Dense<Real>(undef,10)
```
to make a length 10 Dense storage with undefined data. The original constructor `Dense<Real>(10)` still makes storage filled with zeros.

In this branch, the new constructor is used in the Dense*Dense contraction code, which I have seen speeds up Dense tensor contraction. The plan is to use the new constructor in the QDense*QDense contraction code as well.