@mtfishman 2019-11-15
The latest timings on this branch, after changing the QDense storage to allocate uninitialized memory in the contraction code (instead of zeros):
```
    vN Entropy at center bond b=8 = 3.019814122626
    Eigs at center bond b=8: 0.1677 0.1677 0.1597 0.1597 0.0138 0.0138 0.0138 0.0138 0.0096 0.0096 
    Largest link dim during sweep 5/5 was 7341
    Largest truncation error: 4.5789e-09
    Energy after sweep 5/5 is -18.876818002901
    Sweep 5/5 CPU time = 16m, 46s (Wall time = 2m, 0.9s)
```
so around a 20% speedup from the original pull request.