@emstoudenmire 2019-11-15
This is all really great and overdue. I should have done some of this profiling before, but partly it was because I ran into the grpof threading issue (and also I did catch the ldiv one using timers but sort of didn't believe those results or have time to experiment with switching the block representation). 

Once you get the setting of elements and fill handled, I'd say this is ready to go. Will be great to have the same logic going into the Julia version.