@mtfishman 2019-11-19
@emstoudenmire I believe this is ready to go in terms of functionality that needs to be added.

I think the only thing left to think about is the proper way to merge this, since technically it is a breaking change to reading and writing QDense storage.

We haven't been doing a good job tagging minor releases, but I was think that we could tag the current commit of `v3` as `v3.0`, and then when we merge this branch it will be considered `v3.1`. This would help mark when the breaking change will happen, and we can make some functionality to read pre-v3.1 ITensor files.