@emstoudenmire 2019-11-20
Hi Matt, this is great. Glad to see the really big, concrete speedups without much increase, if any, in code complexity.

For the versioning, I think we should be able to go back and tag a v3.0.0 from when we first made the v3 branch (should have done that already! thought we had). Then I think we should tag this one as v3.1.0.